requirements.txt
Django==2.0
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2017.3
six==1.11.0
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-guardian==1.4.9
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2017.3
six==1.11.0
